+++ lib/maru/builder/exceptions.ex
@@ -3,8 +3,6 @@ defmodule Maru.Builder.Exceptions do
   Handle exceptions of current router.
   """
 
-  alias Maru.Router.Endpoint
-
   @doc false
   defmacro rescue_from(:all, [as: error_var], [do: block]) do
     quote do
@@ -36,16 +34,14 @@ defmodule Maru.Builder.Exceptions do
   def make_rescue_block({:all, block}) do
     quote do
       _ ->
-        resp = unquote(block)
-        Endpoint.send_resp(var!(conn), resp)
+        unquote(block)
     end
   end
 
   def make_rescue_block({error, block}) do
     quote do
       unquote(error) ->
-        resp = unquote(block)
-        Endpoint.send_resp(var!(conn), resp)
+        unquote(block)
     end
   end
 
@@ -53,16 +49,14 @@ defmodule Maru.Builder.Exceptions do
   def make_rescue_block({:all, error_var, block}) do
     quote do
       unquote(error_var) ->
-        resp = unquote(block)
-        Endpoint.send_resp(var!(conn), resp)
+        unquote(block)
     end
   end
 
   def make_rescue_block({error, error_var, block}) do
     quote do
       unquote(error_var) in unquote(error) ->
-        resp = unquote(block)
-        Endpoint.send_resp(var!(conn), resp)
+        unquote(block)
     end
   end
 end