+++ lib/maru/helpers/response.ex
@@ -7,6 +7,7 @@ defmodule Maru.Helpers.Response do
   defmacro __using__(_) do
     quote do
       import unquote(__MODULE__)
+      import Plug.Conn
     end
   end
 
@@ -24,6 +25,7 @@ defmodule Maru.Helpers.Response do
   Get assigns.
   """
   defmacro assigns do
+    IO.puts :stderr, "assigns is deprecated in favor of conn.assigns."
     quote do
       var!(conn).assigns
     end
@@ -34,6 +36,7 @@ defmodule Maru.Helpers.Response do
   Set assign.
   """
   defmacro assign(key, value) do
+    IO.puts :stderr, "assign/2 is deprecated in favor of assign/3."
     quote do
       var!(conn) = var!(conn)
    |> Plug.Conn.assign(unquote(key), unquote(value))
@@ -45,6 +48,7 @@ defmodule Maru.Helpers.Response do
   Get headers.
   """
   defmacro headers do
+    IO.puts :stderr, "headers is deprecated in favor of conn.req_headers."
     quote do
       var!(conn).req_headers |> Enum.into %{}
     end
@@ -55,6 +59,7 @@ defmodule Maru.Helpers.Response do
   Set or delete header.
   """
   defmacro header(key, nil) do
+    IO.puts :stderr, "header/2 is deprecated in favor of put_resp_header/3, update_resp_header/3, delete_resp_header/3, merge_resp_headers/3."
     quote do
       var!(conn) = var!(conn)
    |> Plug.Conn.delete_resp_header(unquote(key))
@@ -62,6 +67,7 @@ defmodule Maru.Helpers.Response do
   end
 
   defmacro header(key, value) do
+    IO.puts :stderr, "header/2 is deprecated in favor of put_resp_header/3, update_resp_header/3, delete_resp_header/3, merge_resp_headers/3."
     quote do
       var!(conn) = var!(conn)
    |> Plug.Conn.put_resp_header(unquote(key), unquote(value))
@@ -69,10 +75,9 @@ defmodule Maru.Helpers.Response do
   end
 
 
-  @doc """
-  fetch request body.
-  """
+  @doc false
   defmacro fetch_req_body do
+    IO.puts :stderr, "retch_req_body/1 is deprecated."
     quote do
       {state, body, _} = var!(conn) |> Plug.Conn.read_body
       var!(conn) = var!(conn) |> Plug.Conn.put_private(:maru_body, {state, body})
@@ -80,10 +85,9 @@ defmodule Maru.Helpers.Response do
   end
 
 
-  @doc """
-  Get request body.
-  """
+  @doc false
   defmacro req_body do
+    IO.puts :stderr, "req_body/1 is deprecated."
     quote do
       case var!(conn).private.maru_body do
         nil -> {:error, :unfetched}
@@ -97,6 +101,7 @@ defmodule Maru.Helpers.Response do
   Set content_type.
   """
   defmacro content_type(value) do
+    IO.puts :stderr, "content_type/1 is deprecated in favor of put_resp_content_type/2"
     quote do
       var!(conn) = var!(conn)
    |> Plug.Conn.put_resp_content_type(unquote(value))
@@ -108,6 +113,7 @@ defmodule Maru.Helpers.Response do
   Set status.
   """
   defmacro status(value) do
+    IO.puts :stderr, "status/1 is deprecated in favor of put_status/2"
     quote do
       var!(conn) = var!(conn)
    |> Plug.Conn.put_status(unquote(value))
@@ -119,34 +125,16 @@ defmodule Maru.Helpers.Response do
   Make response by (maru_entity)[https://hex.pm/packages/maru_entity] or key-value pairs.
   """
   defmacro present(payload, opts) do
+    IO.puts :stderr, "present/2 is deprecated in favor of put_present/3"
     quote do
       var!(conn) = var!(conn) |> unquote(__MODULE__).put_present(unquote(payload), unquote(opts))
     end
   end
 
   defmacro present(key, payload, opts) do
+    IO.puts :stderr, "present/3 is deprecated in favor of put_present/4"
     quote do
       var!(conn) = var!(conn) |> unquote(__MODULE__).put_present(unquote(key), unquote(payload), unquote(opts))
-    end
-  end
-
-  def put_present(conn, payload, opts) do
-    value = get_present_value(payload, opts)
-    conn |> Plug.Conn.put_private(:maru_present, value)
-  end
-
-  def put_present(conn, key, payload, opts) do
-    present = conn.private[:maru_present] || %{}
-    value = present |> put_in([key], get_present_value(payload, opts))
-    conn |> Plug.Conn.put_private(:maru_present, value)
-  end
-
-  defp get_present_value(payload, opts) do
-    opts |> Enum.into(%{}) |> Dict.pop(:with) |> case do
-      {nil, _} ->
-        payload
-      {entity_klass, opts} ->
-        entity_klass.serialize(payload, opts)
     end
   end
 
@@ -154,21 +142,34 @@ defmodule Maru.Helpers.Response do
   @doc """
   Make response by url redirect.
   """
-  defmacro redirect(url) do
-    quote do
-      var!(conn) = var!(conn)
-   |> Plug.Conn.put_resp_header("location", unquote(url))
-   |> Plug.Conn.put_status(302)
-   |> Plug.Conn.halt
-    end
+  def redirect(%Plug.Conn{}=conn, url, opts \\ []) do
+    code = opts[:permanent] && 301 || 302
+    conn
+ |> Plug.Conn.put_resp_header("location", url)
+ |> Plug.Conn.put_status(code)
+ |> Plug.Conn.halt
   end
 
-  defmacro redirect(url, permanent: true) do
-    quote do
-      var!(conn) = var!(conn)
-   |> Plug.Conn.put_resp_header("location", unquote(url))
-   |> Plug.Conn.put_status(301)
-   |> Plug.Conn.halt
-    end
+
+  def json(%Plug.Conn{}=conn, data) do
+    conn
+ |> Plug.Conn.put_resp_content_type("application/json")
+ |> Plug.Conn.send_resp(conn.status || 200, Poison.encode_to_iodata!(data))
+ |> Plug.Conn.halt
   end
+
+  def html(%Plug.Conn{}=conn, data) do
+    conn
+ |> Plug.Conn.put_resp_content_type("text/html")
+ |> Plug.Conn.send_resp(conn.status || 200, to_string(data))
+ |> Plug.Conn.halt
+  end
+
+  def text(%Plug.Conn{}=conn, data) do
+    conn
+ |> Plug.Conn.put_resp_content_type("text/plain")
+ |> Plug.Conn.send_resp(conn.status || 200, to_string(data))
+ |> Plug.Conn.halt
+  end
+
 end